+++ internal/storage/sql/common/evaluation.go
@@ -5,8 +5,8 @@ import (
 	"database/sql"
 
 	sq "github.com/Masterminds/squirrel"
+	"go.flipt.io/flipt/internal/storage"
 	flipt "go.flipt.io/flipt/rpc/flipt"
-	"go.flipt.io/flipt/storage"
 )
 
 type optionalConstraint struct {
