+++ internal/storage/sql/common/rule.go
@@ -12,8 +12,8 @@ import (
 	"google.golang.org/protobuf/types/known/timestamppb"
 
 	errs "go.flipt.io/flipt/errors"
+	"go.flipt.io/flipt/internal/storage"
 	flipt "go.flipt.io/flipt/rpc/flipt"
-	"go.flipt.io/flipt/storage"
 )
 
 // GetRule gets an individual rule
