+++ internal/storage/sql/postgres/postgres.go
@@ -1,4 +1,4 @@
-package mysql
+package postgres
 
 import (
 	"context"
@@ -7,23 +7,23 @@ import (
 	"errors"
 
 	sq "github.com/Masterminds/squirrel"
-	"github.com/go-sql-driver/mysql"
+	"github.com/lib/pq"
 	errs "go.flipt.io/flipt/errors"
+	"go.flipt.io/flipt/internal/storage"
+	"go.flipt.io/flipt/internal/storage/sql/common"
 	flipt "go.flipt.io/flipt/rpc/flipt"
-	"go.flipt.io/flipt/storage"
-	"go.flipt.io/flipt/storage/sql/common"
 	"go.uber.org/zap"
 )
 
 const (
-	constraintForeignKeyErrCode uint16 = 1452
-	constraintUniqueErrCode     uint16 = 1062
+	constraintForeignKeyErr = "foreign_key_violation"
+	constraintUniqueErr     = "unique_violation"
 )
 
 var _ storage.Store = &Store{}
 
 func NewStore(db *sql.DB, logger *zap.Logger) *Store {
-	builder := sq.StatementBuilder.RunWith(sq.NewStmtCacher(db))
+	builder := sq.StatementBuilder.PlaceholderFormat(sq.Dollar).RunWith(sq.NewStmtCacher(db))
 
 	return &Store{
 		Store: common.NewStore(db, builder, logger),
@@ -35,16 +35,16 @@ type Store struct {
 }
 
 func (s *Store) String() string {
-	return "mysql"
+	return "postgres"
 }
 
 func (s *Store) CreateFlag(ctx context.Context, r *flipt.CreateFlagRequest) (*flipt.Flag, error) {
 	flag, err := s.Store.CreateFlag(ctx, r)
 
 	if err != nil {
-		var merr *mysql.MySQLError
+		var perr *pq.Error
 
-		if errors.As(err, &merr) && merr.Number == constraintUniqueErrCode {
+		if errors.As(err, &perr) && perr.Code.Name() == constraintUniqueErr {
 			return nil, errs.ErrInvalidf("flag %q is not unique", r.Key)
 		}
 
@@ -58,13 +58,13 @@ func (s *Store) CreateVariant(ctx context.Context, r *flipt.CreateVariantRequest
 	variant, err := s.Store.CreateVariant(ctx, r)
 
 	if err != nil {
-		var merr *mysql.MySQLError
+		var perr *pq.Error
 
-		if errors.As(err, &merr) {
-			switch merr.Number {
-			case constraintForeignKeyErrCode:
+		if errors.As(err, &perr) {
+			switch perr.Code.Name() {
+			case constraintForeignKeyErr:
 				return nil, errs.ErrNotFoundf("flag %q", r.FlagKey)
-			case constraintUniqueErrCode:
+			case constraintUniqueErr:
 				return nil, errs.ErrInvalidf("variant %q is not unique", r.Key)
 			}
 		}
@@ -79,9 +79,9 @@ func (s *Store) UpdateVariant(ctx context.Context, r *flipt.UpdateVariantRequest
 	variant, err := s.Store.UpdateVariant(ctx, r)
 
 	if err != nil {
-		var merr *mysql.MySQLError
+		var perr *pq.Error
 
-		if errors.As(err, &merr) && merr.Number == constraintUniqueErrCode {
+		if errors.As(err, &perr) && perr.Code.Name() == constraintUniqueErr {
 			return nil, errs.ErrInvalidf("variant %q is not unique", r.Key)
 		}
 
@@ -95,9 +95,9 @@ func (s *Store) CreateSegment(ctx context.Context, r *flipt.CreateSegmentRequest
 	segment, err := s.Store.CreateSegment(ctx, r)
 
 	if err != nil {
-		var merr *mysql.MySQLError
+		var perr *pq.Error
 
-		if errors.As(err, &merr) && merr.Number == constraintUniqueErrCode {
+		if errors.As(err, &perr) && perr.Code.Name() == constraintUniqueErr {
 			return nil, errs.ErrInvalidf("segment %q is not unique", r.Key)
 		}
 
@@ -111,9 +111,9 @@ func (s *Store) CreateConstraint(ctx context.Context, r *flipt.CreateConstraintR
 	constraint, err := s.Store.CreateConstraint(ctx, r)
 
 	if err != nil {
-		var merr *mysql.MySQLError
+		var perr *pq.Error
 
-		if errors.As(err, &merr) && merr.Number == constraintForeignKeyErrCode {
+		if errors.As(err, &perr) && perr.Code.Name() == constraintForeignKeyErr {
 			return nil, errs.ErrNotFoundf("segment %q", r.SegmentKey)
 		}
 
@@ -127,9 +127,9 @@ func (s *Store) CreateRule(ctx context.Context, r *flipt.CreateRuleRequest) (*fl
 	rule, err := s.Store.CreateRule(ctx, r)
 
 	if err != nil {
-		var merr *mysql.MySQLError
+		var perr *pq.Error
 
-		if errors.As(err, &merr) && merr.Number == constraintForeignKeyErrCode {
+		if errors.As(err, &perr) && perr.Code.Name() == constraintForeignKeyErr {
 			return nil, errs.ErrNotFoundf("flag %q or segment %q", r.FlagKey, r.SegmentKey)
 		}
 
@@ -143,9 +143,9 @@ func (s *Store) CreateDistribution(ctx context.Context, r *flipt.CreateDistribut
 	dist, err := s.Store.CreateDistribution(ctx, r)
 
 	if err != nil {
-		var merr *mysql.MySQLError
+		var perr *pq.Error
 
-		if errors.As(err, &merr) && merr.Number == constraintForeignKeyErrCode {
+		if errors.As(err, &perr) && perr.Code.Name() == constraintForeignKeyErr {
 			return nil, errs.ErrNotFoundf("rule %q", r.RuleId)
 		}
