+++ internal/storage/sql/sqlite/sqlite.go
@@ -1,4 +1,4 @@
-package postgres
+package sqlite
 
 import (
 	"context"
@@ -7,44 +7,41 @@ import (
 	"errors"
 
 	sq "github.com/Masterminds/squirrel"
-	"github.com/lib/pq"
+	"github.com/mattn/go-sqlite3"
 	errs "go.flipt.io/flipt/errors"
+	"go.flipt.io/flipt/internal/storage"
+	"go.flipt.io/flipt/internal/storage/sql/common"
 	flipt "go.flipt.io/flipt/rpc/flipt"
-	"go.flipt.io/flipt/storage"
-	"go.flipt.io/flipt/storage/sql/common"
 	"go.uber.org/zap"
-)
-
-const (
-	constraintForeignKeyErr = "foreign_key_violation"
-	constraintUniqueErr     = "unique_violation"
 )
 
 var _ storage.Store = &Store{}
 
+// NewStore creates a new sqlite.Store
 func NewStore(db *sql.DB, logger *zap.Logger) *Store {
-	builder := sq.StatementBuilder.PlaceholderFormat(sq.Dollar).RunWith(sq.NewStmtCacher(db))
+	builder := sq.StatementBuilder.RunWith(sq.NewStmtCacher(db))
 
 	return &Store{
 		Store: common.NewStore(db, builder, logger),
 	}
 }
 
+// Store is a sqlite specific implementation of storage.Store
 type Store struct {
 	*common.Store
 }
 
 func (s *Store) String() string {
-	return "postgres"
+	return "sqlite"
 }
 
 func (s *Store) CreateFlag(ctx context.Context, r *flipt.CreateFlagRequest) (*flipt.Flag, error) {
 	flag, err := s.Store.CreateFlag(ctx, r)
 
 	if err != nil {
-		var perr *pq.Error
+		var serr sqlite3.Error
 
-		if errors.As(err, &perr) && perr.Code.Name() == constraintUniqueErr {
+		if errors.As(err, &serr) && serr.Code == sqlite3.ErrConstraint {
 			return nil, errs.ErrInvalidf("flag %q is not unique", r.Key)
 		}
 
@@ -58,13 +55,13 @@ func (s *Store) CreateVariant(ctx context.Context, r *flipt.CreateVariantRequest
 	variant, err := s.Store.CreateVariant(ctx, r)
 
 	if err != nil {
-		var perr *pq.Error
+		var serr sqlite3.Error
 
-		if errors.As(err, &perr) {
-			switch perr.Code.Name() {
-			case constraintForeignKeyErr:
+		if errors.As(err, &serr) {
+			switch serr.ExtendedCode {
+			case sqlite3.ErrConstraintForeignKey:
 				return nil, errs.ErrNotFoundf("flag %q", r.FlagKey)
-			case constraintUniqueErr:
+			case sqlite3.ErrConstraintUnique:
 				return nil, errs.ErrInvalidf("variant %q is not unique", r.Key)
 			}
 		}
@@ -79,9 +76,9 @@ func (s *Store) UpdateVariant(ctx context.Context, r *flipt.UpdateVariantRequest
 	variant, err := s.Store.UpdateVariant(ctx, r)
 
 	if err != nil {
-		var perr *pq.Error
+		var serr sqlite3.Error
 
-		if errors.As(err, &perr) && perr.Code.Name() == constraintUniqueErr {
+		if errors.As(err, &serr) && serr.Code == sqlite3.ErrConstraint {
 			return nil, errs.ErrInvalidf("variant %q is not unique", r.Key)
 		}
 
@@ -95,9 +92,9 @@ func (s *Store) CreateSegment(ctx context.Context, r *flipt.CreateSegmentRequest
 	segment, err := s.Store.CreateSegment(ctx, r)
 
 	if err != nil {
-		var perr *pq.Error
+		var serr sqlite3.Error
 
-		if errors.As(err, &perr) && perr.Code.Name() == constraintUniqueErr {
+		if errors.As(err, &serr) && serr.Code == sqlite3.ErrConstraint {
 			return nil, errs.ErrInvalidf("segment %q is not unique", r.Key)
 		}
 
@@ -111,9 +108,9 @@ func (s *Store) CreateConstraint(ctx context.Context, r *flipt.CreateConstraintR
 	constraint, err := s.Store.CreateConstraint(ctx, r)
 
 	if err != nil {
-		var perr *pq.Error
+		var serr sqlite3.Error
 
-		if errors.As(err, &perr) && perr.Code.Name() == constraintForeignKeyErr {
+		if errors.As(err, &serr) && serr.Code == sqlite3.ErrConstraint {
 			return nil, errs.ErrNotFoundf("segment %q", r.SegmentKey)
 		}
 
@@ -127,9 +124,9 @@ func (s *Store) CreateRule(ctx context.Context, r *flipt.CreateRuleRequest) (*fl
 	rule, err := s.Store.CreateRule(ctx, r)
 
 	if err != nil {
-		var perr *pq.Error
+		var serr sqlite3.Error
 
-		if errors.As(err, &perr) && perr.Code.Name() == constraintForeignKeyErr {
+		if errors.As(err, &serr) && serr.Code == sqlite3.ErrConstraint {
 			return nil, errs.ErrNotFoundf("flag %q or segment %q", r.FlagKey, r.SegmentKey)
 		}
 
@@ -143,9 +140,9 @@ func (s *Store) CreateDistribution(ctx context.Context, r *flipt.CreateDistribut
 	dist, err := s.Store.CreateDistribution(ctx, r)
 
 	if err != nil {
-		var perr *pq.Error
+		var serr sqlite3.Error
 
-		if errors.As(err, &perr) && perr.Code.Name() == constraintForeignKeyErr {
+		if errors.As(err, &serr) && serr.Code == sqlite3.ErrConstraint {
 			return nil, errs.ErrNotFoundf("rule %q", r.RuleId)
 		}
