+++ server/evaluator.go
@@ -9,8 +9,8 @@ import (
 	"strings"
 
 	errs "go.flipt.io/flipt/errors"
+	"go.flipt.io/flipt/internal/storage"
 	flipt "go.flipt.io/flipt/rpc/flipt"
-	"go.flipt.io/flipt/storage"
 	"go.uber.org/zap"
 )
